13-April-2003 Akiya Miyamoto   Modified so as not to use Imakefile   Modification to use with JAVA is also included

+++ util/fortlib/src/nuaddr.c
@@ -16,6 +16,6 @@
 
 nuaddr_(x) int *x ;{
 	int iadr ;
-	iadr = x ;
+	iadr = (int)x ;
 	return(iadr) ;
 }
